apply coderabbit code enhancements
