+++ src/app/coc3/page.tsx
@@ -367,4 +367,4 @@ export default function COC3Page() {
       <Footer />
     </div>
   )
-}
+}
